tt-comfy-bridge/requirements.txt
msgpack>=1.0.0
torch>=2.0.0
numpy>=1.24.0
aiofiles>=23.0.0
Pillow>=10.0.0

